refactor: waiter (#20)
The wait function was not returning the latest asset and we were not
updating the asset in state.  This is a problem because when an asset
goes from `REQUESTED` to `COMPLETED` there could be different `outputs`
which are required for other asset resources.

Furthermore, we were not differentiating between `FAILED` and `COMPLETED`
statuses when waiting for an asset.  This means that an asset could be
failed to be created but terraform continues to the next step in its
plan.

We also added a `make resource` command to copy `acm/resources.go` to all
other aws asset packages.  This simply copy and pastes acm/resources to
all other aws assets which will help us make changes across all assets
until we come up with a better solution.

I also fixed a type conversion error for redis and rds when converting
an asset output to a plan object.