feat: user no longer provides asset version
The user doesn't need to know what an asset version is or what it is
used for at this point in time and only serves to confuse people.
Instead we converted `asset_version` to a computed property that will
still be stored in state but not required since we always want to
provision an asset with the `latest` asset version.